refactor: training script and dataset preparetion
